Five references that were not holding what the program holds
Four of these move the column in our favour and one moves it against us, and
none of the five is a code generation change. A reader discounting today's
numbers should discount exactly this much of them.

The rule: a reference is what a competent Java programmer would write, and must
decline anything that puts a cost in one lane only -- no boxing where the
subject boxes nothing, no field narrower than the f64 a TypeScript `number` is.
Five references broke the second clause or the first.

  array-predicates    int[]              -> double[]      3.28x -> 1.71x
  array-methods       int[]              -> double[]      1.67x -> 1.18x
  arrays              int[], int total   -> double[]      1.35x -> 1.02x
  array-from          int[], Set<Integer>-> double[], Set<Double>   2.39x -> 1.95x
  array-mutations     ArrayList<Integer> -> a double[] list         1.06x -> 0.68x

Each of the four narrow ones carried a comment arguing at length that the
narrower reference was the deliberate harder one and that our `double[]` was a
real gap it should not hide. That argument is wrong about where the gap belongs:
element width is a middle-end representation choice, not something this backend
decides, and putting it in the ratio meant the ratio stopped being about code
generation. It is filed with `hir` with a number instead.

I also had it wrong about the size. Record 0147 published element width at 1.6%,
from instructions per operation -- the instrument I had chosen precisely because
this row's timing varies 1.22x between runs. It is the wrong instrument for a
width question, because width does not change how many instructions execute:

  int[]      33,139 instr/op    6,969 cycles/op    IPC 4.76
  double[]   35,717             13,293             IPC 2.69
             +7.8%              +91%

The row stalls rather than issues. `some`, `every` and `findIndex` each break on
their own comparison, so the compare's latency is the loop-carried dependence
and `dcmpl` is several cycles where `if_icmplt` is one. 0147 carries the
correction rather than the edit.

`array-mutations` went the other way and refutes the reasoning usually given for
the no-boxing rule. Replacing `ArrayList<Integer>` with a hand-written growable
`double[]` made the reference *slower* -- 2.03 us to 2.48 us -- because an
`Object[]` of small `Integer`s moves four bytes an element under compressed oops
where a `double[]` moves eight, and `ArrayList`'s shift and splice are
`System.arraycopy` over that. The rule is still right; the row is one we win by
less than the old number said rather than more.

Not changed, and worth saying why: `generic-classes` keeps its `Box<Integer>`.
Java generics are erased and there is no non-boxing generic to write -- a
hand-specialised box would be a different program -- and the row sits at 1.03x
against it, which is the evidence that C2 removes those boxes rather than an
argument that it should.

All five agree with node by bit pattern; the harness's per-family work guard
compared them.

+++ benches/cases/array-from/ref.java
@@ -1,32 +1,28 @@
-// What a Java programmer writes for `Array.from`: `Arrays.copyOf` for an array
-// and `toArray` for a set.
+// What a Java programmer writes for `Array.from` over an array and over a
+// `Set`: `Arrays.copyOf` and `Set.toArray`.
 //
-// Two different copies, which is why the case names both. `Arrays.copyOf` is an
-// intrinsic that becomes a bulk memory move; `Set::toArray` has to walk the
-// table and cannot, so it is a scan with a store per element. The TypeScript's
-// two `Array.from` calls have exactly that split under them, and the row is
-// whether ours splits the same way.
+// `Array.from(xs)` on an array is a copy and nothing else, which is what
+// `Arrays.copyOf` is. `Array.from(set)` walks the set and materialises it,
+// which is `toArray` -- and `toArray` on a `HashSet` allocates an `Object[]`
+// and fills it by iterating, exactly as the lowering does.
 //
-// The set's *length* is read rather than an element, for the reason the
-// TypeScript gives: a `Set`'s iteration order is insertion order in JavaScript
-// and hash order in Java, so indexing one would compare two orders rather than
-// two copies. The walk runs to the end either way, so the work is the same and
-// only the comparison is made safe.
+// **`double[]` and `Set<Double>`, not `int[]` and `Set<Integer>`.** The
+// narrower pair was here deliberately, as the harder reference; it is the
+// thing the rule forbids, and on a row that is mostly `Arrays.copyOf` of 256
+// elements the width is half the bytes moved. `Set<Integer>` is worse than
+// narrow -- `Integer.valueOf` caches -128..127 and `Double.valueOf` caches
+// nothing, so half the keys here were free on one lane and allocated on the
+// other. Our own `NtsMap` keys through `Double.valueOf`, so `Set<Double>` is
+// the matched representation rather than the generous one.
 //
-// `Set<Integer>` boxes, as `map-and-set/ref.java` does and for the same reason:
-// the JDK has no primitive set.
+// Correcting it moved the published row from 2.39x to 1.95x, and **none of
+// that is code generation.** What is left is `Arrays.copyOf` of 256 elements,
+// two thousand times, which is the thing the row is for.
 //
-// `int[]` rather than `double[]`, which is the deliberate harder reference
-// `arrays` and `array-predicates` also write, and it is worth naming here
-// because on this row it is most of the number. The TypeScript is `number[]`,
-// so the lane prepares `managed<[f64]>` and copies eight bytes an element where
-// this copies four. Measured rather than assumed: **8,280,888 bytes/op against
-// the reference's 4,176,848**, a factor of 1.98 on a row whose whole subject is
-// two copies.
-//
-// So the gap is element width and the row should show it. `hir::elements`
-// proving these are int32 is what closes it, and `dispatch` -- which already
-// prepares as `[i32]` -- is what that looks like when it has happened.
+// The *length* of the listed set, not an element: a `Set`'s iteration order is
+// insertion order for us and a hash order for `HashSet`, so indexing one would
+// compare two orders rather than two copies. The walk runs to the end either
+// way, which is the part being measured.
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
@@ -36,18 +32,18 @@ final class Ref extends Bench.Work {
     private static volatile double seed = 5;
 
     static double work(int seed) {
-        int[] xs = new int[256];
+        double[] xs = new double[256];
         for (int i = 0; i < 256; i++) {
             xs[i] = i + seed;
         }
-        Set<Integer> marks = new HashSet<>();
+        Set<Double> marks = new HashSet<>();
         for (int i = 0; i < 256; i++) {
-            marks.add(i * 3 + seed);
+            marks.add((double) (i * 3 + seed));
         }
 
         double total = 0;
         for (int round = 0; round < 2000; round++) {
-            int[] copied = Arrays.copyOf(xs, xs.length);
+            double[] copied = Arrays.copyOf(xs, xs.length);
             total = total + copied[round % 256];
             Object[] listed = marks.toArray();
             total = total + listed.length;
